Model object, ready for sampling

Model object, ready for sampling

Model object, ready for sampling

See also

Public fields

form

Form of the neuromodulation (additive or multiplicative)

Methods

Public methods

Model$new()

Model$sample()

Model$make_standata()

Model$clone()

Method new()

Create a new Model object.

Usage

Model$new(form)

Arguments

form

Form of neuromodulation in the model.

Returns

A new Model object.

Method sample()

Draw samples from the posterior of the model

Usage

Model$sample(d = NULL, standata = NULL, prior = Prior$new(), ...)

Arguments

d

dataframe from which to make standata

standata

data to pass directly to stan. Can be produced with make_standata(). If present, this data will be used rather than the data stored in d.

prior

An object of class Prior

...

further arguments passed to rstan::sampling().

Returns

A new ModelFit. Prepare data for running model

Method make_standata()

Usage

Model$make_standata(d, prior = Prior$new())

Arguments

d

dataframe from which to make standata. The output of this can be passed to vtf().

prior

A Prior

Returns

named list

Method clone()

The objects of this class are cloneable with this method.

Usage

Model$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.