ModelFit object, output from sampling

ModelFit object, output from sampling

ModelFit object, output from sampling

Public fields

rawdata

The original data that was used to prepare the model

prior

The Prior that was used to fit the model

stanfit

rstan::stanfit output

model

Model object used to generate the fit

Methods

Public methods

ModelFit$new()

ModelFit$standata()

ModelFit$loo()

ModelFit$make_vtf0()

ModelFit$.lfun()

ModelFit$.llfun()

ModelFit$clone()

Method new()

Create a new ModelFit

Usage

ModelFit$new(rawdata, prior, model, stanfit)

Arguments

rawdata

dataframe from which to make standata

prior

Prior that was used to fit the model

model

Model used to draw samples

stanfit

rstan::stanfit output by model

Returns

A new ModelFit object.

Method standata()

Reconstruct the list of data that was passed to Stan

Usage

ModelFit$standata()

Returns

A named list Estimate the LOO score

Method loo()

Usage

ModelFit$loo(cores = 1)

Arguments

cores

int

Returns

A loo::psis_loo object create voxel tuning function posterior

Method make_vtf0()

Usage

ModelFit$make_vtf0(cores = 1)

Arguments

cores

integer number of cores over which to parallelize

Returns

array n_draws x n_chains x n_unique obs. for use in ModelFit$loo() Likelihood function

Method .lfun()

Usage

ModelFit$.lfun(data_i, draws)

Arguments

data_i

single row of data

draws

posterior

Returns

The likelihood of a single observation Log-Likelihood function

Method .llfun()

Usage

ModelFit$.llfun(data_i, draws)

Arguments

data_i

single row of data

draws

posterior

Returns

The log-likelihood of a single observation

Method clone()

The objects of this class are cloneable with this method.

Usage

ModelFit$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.