Sample beta values

Example dataset for testing analyses, and to see input structure

sub02

Format

Objects of class tbl_df (inherits from tbl, data.frame) a minimal tibble for the vtuner analysis (53856 x 7)

sub

factor ID of subject

run

factor ID of run

voxel

factor ID of voxel

contrast

Factor with different values of contrast. Note that the factor must have the lower levels of contrast correspond to a lower integer representation (e.g., low == 1 and high == 2)

orientation

Orientation presented on trial (-pi,pi)

y

beta values in given condition (corresponds to beta in betas2)

ses

factor id of session in which run was collected

Value

a tibble

Examples

sub02

#> # A tibble: 53,856 x 5
#>    sub   voxel  contrast orientation      y
#>    <fct> <fct>  <fct>          <dbl>  <dbl>
#>  1 2     191852 low            0.785  3.36 
#>  2 2     197706 low            0.785 -2.84 
#>  3 2     197769 low            0.785 -2.03 
#>  4 2     197842 low            0.785  2.23 
#>  5 2     197906 low            0.785  2.86 
#>  6 2     197907 low            0.785  1.51 
#>  7 2     203688 low            0.785 -2.75 
#>  8 2     203689 low            0.785 -2.51 
#>  9 2     203690 low            0.785 -3.16 
#> 10 2     203753 low            0.785  0.711
#> # … with 53,846 more rows